DDD e primeiro framework-nest/src/domain/notification/application/use-cases/read-notification.ts
import { Either, left, right } from '@/core/either';
import { Notification } from '@/domain/notification/enterprise/entities/notification';
import { INotificationsRepository } from '@/domain/notification/application/repositories/notifications-repository';
import { ResourceNotFoundError } from '@/core/errors/errors/resource-not-found-error';
import { NotAllowedError } from '@/core/errors/errors/not-allowed-error';
import { Injectable } from '@nestjs/common';

interface IReadNotificationUseCaseParams {
  recipientId: string;
  notificationId: string;
}

type ReadNotificationUseCaseResponse = Either<
  ResourceNotFoundError | NotAllowedError,
  {
    notification: Notification;
  }
>;

@Injectable()
export class ReadNotificationUseCase {
  constructor(private notificationsRepository: INotificationsRepository) {}

  async handle({
    recipientId,
    notificationId,
  }: IReadNotificationUseCaseParams): Promise<ReadNotificationUseCaseResponse> {
    const notification = await this.notificationsRepository.findById(notificationId);

    if (!notification) {
      return left(new ResourceNotFoundError());
    }

    if (recipientId !== notification.recipientId.toString()) {
      return left(new NotAllowedError());
    }

    notification.read();

    await this.notificationsRepository.save(notification);

    return right({
      notification,
    });
  }
}
